Removed unnecessary files from repo.  Updated files in quality of LQUBO and methods to use new form_LQUBO class.  Removed all older variatons of form_LQUBO.
