+++ results/convergence/tsp_HD_Slice_LQUBO_20.csv
@@ -1,0 +1,21 @@
+,domain,convergence percent error vals
+0,5,1.2025398530199347
+1,10,1.1332899748407608
+2,15,1.0991476264456919
+3,20,1.0756478853131428
+4,25,1.0616479552920246
+5,30,1.040060998490454
+6,35,1.0268728034459413
+7,40,1.0177738694012664
+8,45,1.010009753080193
+9,50,1.0042987344434897
+10,55,0.9991571051983664
+11,60,0.9949145265265883
+12,65,0.9877530546001452
+13,70,0.9803264833393946
+14,75,0.9756560017964452
+15,80,0.9713138089716677
+16,85,0.9687615154953095
+17,90,0.9647672307908028
+18,95,0.9610844102941845
+19,100,0.956994743862281
