+++ results/convergence/tsp_HD_Slice_LQUBO_WS_20.csv
@@ -1,0 +1,21 @@
+,domain,convergence percent error vals
+0,5,1.0967058229698827
+1,10,1.0266112175150952
+2,15,0.983947233751754
+3,20,0.9545356075507843
+4,25,0.9319961505999105
+5,30,0.9214222213907413
+6,35,0.9106623777903677
+7,40,0.8962516269503003
+8,45,0.8883887828378387
+9,50,0.8798415709722704
+10,55,0.8709207528083662
+11,60,0.8625941701641241
+12,65,0.8557524985695325
+13,70,0.8528907675657312
+14,75,0.8496380114356921
+15,80,0.8446495219539212
+16,85,0.8393111554933715
+17,90,0.8359557605870267
+18,95,0.8319524949920886
+19,100,0.8304508157695445
