+++ results/convergence/tsp_LQUBO_20.csv
@@ -1,0 +1,21 @@
+,domain,convergence percent error vals
+0,5,1.2171129321334697
+1,10,1.1379950799245022
+2,15,1.0997568289851731
+3,20,1.0785305135188104
+4,25,1.060765675413013
+5,30,1.040314523519296
+6,35,1.020859396544963
+7,40,1.0067564819648467
+8,45,0.9999796165494452
+9,50,0.994054904454093
+10,55,0.986225753336179
+11,60,0.9796438905465109
+12,65,0.9739123102227256
+13,70,0.9691789401200724
+14,75,0.9642889189479212
+15,80,0.958774501190097
+16,85,0.9559479516063574
+17,90,0.9536729475312284
+18,95,0.9509497991811213
+19,100,0.944323653718207
