+++ results/convergence/tsp_LQUBO_WP_20.csv
@@ -1,0 +1,7 @@
+,domain,convergence percent error vals
+0,5,0.45013901715388144
+1,10,0.22218731666922395
+2,15,0.22218731666922395
+3,20,0.22218731666922395
+4,25,0.22218731666922395
+5,30,0.22218731666922395
