+++ results/convergence/tsp_LQUBO_WP_and_WS_20.csv
@@ -1,0 +1,21 @@
+,domain,convergence percent error vals
+0,5,1.1759652582804265
+1,10,1.0966401895313238
+2,15,1.048727236179468
+3,20,1.0154568891190623
+4,25,1.002513163545154
+5,30,0.984894896145551
+6,35,0.9755327925507462
+7,40,0.9658612978528877
+8,45,0.9585490351441327
+9,50,0.9505896369169463
+10,55,0.9433282500980852
+11,60,0.9371645444149768
+12,65,0.9324499542765659
+13,70,0.9279000732993788
+14,75,0.9238589507177946
+15,80,0.9215414817741674
+16,85,0.9168837081760539
+17,90,0.909637575979722
+18,95,0.9057016710819992
+19,100,0.901041570375465
