+++ results/convergence/tsp_LQUBO_WS_20.csv
@@ -1,0 +1,21 @@
+,domain,convergence percent error vals
+0,5,1.1108620548091082
+1,10,1.03558281597629
+2,15,0.9990746433361205
+3,20,0.9668975476343974
+4,25,0.9445104593148083
+5,30,0.9287272218701088
+6,35,0.9211236005809127
+7,40,0.910782847231086
+8,45,0.9033274204385686
+9,50,0.8972204837943949
+10,55,0.8909018599232321
+11,60,0.8845240329626761
+12,65,0.8804933527921764
+13,70,0.8766782489410055
+14,75,0.872769121137067
+15,80,0.8689993853758362
+16,85,0.8631509911418266
+17,90,0.8573768426548346
+18,95,0.8535106911380889
+19,100,0.8516543072044738
