+++ results/convergence/tsp_Rand_Slice_LQUBO_20.csv
@@ -1,0 +1,21 @@
+,domain,convergence percent error vals
+0,5,1.2187540403886536
+1,10,1.1473676513137308
+2,15,1.108501741441692
+3,20,1.0744736143031484
+4,25,1.0551005849109367
+5,30,1.0389557560040992
+6,35,1.0245312992945736
+7,40,1.0174623797254034
+8,45,1.0094462358384173
+9,50,1.0006668826478113
+10,55,0.9938519748245017
+11,60,0.9871351221554631
+12,65,0.9758173685713675
+13,70,0.9684765953266404
+14,75,0.9635245636837991
+15,80,0.9571155717626921
+16,85,0.9519333265012392
+17,90,0.946121857760668
+18,95,0.9426412239736464
+19,100,0.9400421913946679
