+++ results/convergence/tsp_Rand_Slice_LQUBO_WS_20.csv
@@ -1,0 +1,21 @@
+,domain,convergence percent error vals
+0,5,1.2072741161843061
+1,10,1.1250005506896372
+2,15,1.0832047070767445
+3,20,1.0568889251743445
+4,25,1.0409544558384258
+5,30,1.0268878167447675
+6,35,1.0121269571904499
+7,40,1.0011320971794972
+8,45,0.9906879222221804
+9,50,0.9822421131238643
+10,55,0.9772130623022008
+11,60,0.9699700111980288
+12,65,0.9647179068444567
+13,70,0.9575760003793449
+14,75,0.9546148120998434
+15,80,0.9504648629001511
+16,85,0.9466018304293072
+17,90,0.9427673945726105
+18,95,0.9411697776414076
+19,100,0.939224482147483
